Add polyfill for Object.assign
This commit polyfills `Object.assign` serverside for both TypeScript-
and Babel-based builds.

Part of the goal here is to show developers how to polyfill methods
that are used by imported libraries, but that are not supported by
Nashorn.

This is why we chose to add new file in resources named
`polyfills.js`, instead of using a npm-library. Resolving the
function should be obvious by looking at the naming of the files,
the export, and the usage in `webpack.server.config.js`.

The polyfill for `Object.assign` has been copied from *starter-esbuild*.
https://github.com/enonic/starter-esbuild/blob/master/src/main/resources/lib/nashorn/assign.js

Closes: #49